+++ src/bltGrIsoline.c
@@ -1240,7 +1240,7 @@ StepsOp(ClientData clientData, Tcl_Interp *interp, int objc,
 {
     Graph *graphPtr = clientData;
     Isoline *isoPtr;
-    long numSteps, i;
+    size_t numSteps, i;
 
     if (Blt_GetCountFromObj(interp, objv[3], COUNT_POS, &numSteps) != TCL_OK) {
         return TCL_ERROR;
