+++ src/bltSwitch.c
@@ -461,7 +461,7 @@ DoSwitch(
 
         case BLT_SWITCH_INT_NNEG:
             {
-                long value;
+                size_t value;
                 
                 if (Blt_GetCountFromObj(interp, objPtr, COUNT_NNEG, 
                         &value) != TCL_OK) {
@@ -473,7 +473,7 @@ DoSwitch(
 
         case BLT_SWITCH_INT_POS:
             {
-                long value;
+                size_t value;
                 
                 if (Blt_GetCountFromObj(interp, objPtr, COUNT_POS, 
                         &value) != TCL_OK) {
@@ -532,7 +532,7 @@ DoSwitch(
 
         case BLT_SWITCH_LONG_NNEG:
             {
-                long value;
+                size_t value;
                 
                 if (Blt_GetCountFromObj(interp, objPtr, COUNT_NNEG, 
                         &value) != TCL_OK) {
@@ -544,7 +544,7 @@ DoSwitch(
 
         case BLT_SWITCH_LONG_POS:
             {
-                long value;
+                size_t value;
                 
                 if (Blt_GetCountFromObj(interp, objPtr, COUNT_POS, &value)
                         != TCL_OK) {
